muon plot prototype
